Changes required by upstream file reorganization
