+++ src/.gitignore
@@ -1,3 +1,4 @@
+/tmp/
 config.json
 credentials.json
 node_modules
