+++ src/MIMS/MbrAddresses-index.js
@@ -1,4 +1,4 @@
-const { config } = require("./getConfig.js");
+const { config } = require("./config/getConfig.js");
 db = db.getSiblingDB( config.wing );
 db.MbrAddresses.createIndex( { CAPID : 1 } );
