+++ src/MIMS/MbrContact-index.js
@@ -1,4 +1,4 @@
-const { config } = require("./getConfig.js");
+const { config } = require("./config/getConfig.js");
 db = db.getSiblingDB( config.wing );
 db.MbrContact.createIndex( { CAPID : 1 } );
