+++ src/MIMS/indexGoogle.js
@@ -1,4 +1,4 @@
-const { config } = require("./getConfig.js");
+const { config } = require("./config/getConfig.js");
 db = db.getSiblingDB( config.wing );
 db.Google.createIndex( { 'primaryEmail' : 1 } );
 db.Google.createIndex( { "name.familyName" : 1 } );
